🚚Rename and move : Route component moved from component directory ♻️Refactoring : Signup Component directory divide Login and view
