+++ src/routes/index.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { Switch, Route } from "react-router-dom/cjs/react-router-dom.min";
-import Home from "../components/Home/Home";
-import Signup from "../components/Signup/Signup";
+import Home from "./Home/Home";
+import Signup from "./Signup/Container";
 import Main from "../components/Main/Main";
 
 function Routes({ isLogin }) {
